mobile-ESP-BLE-init
The mobile can connect to ESP32 by APP through BLE, however, the WiFi info and GPS data was not trans yet. Need further development on both ESP32(Arduino) and APP(Flutter). There are several issues with the APP: After connecting to the device, there is no disconnect button designed, which will cause the IoT device to remain connected to the phone unless one of the devices is disconnected. In addition, after selecting a device, the connection button does not display different content according to the connection status of the corresponding device